Add a warning for SIEC as the default curve
As of right now, SIEC hasn't received third-party cryptanalysis. For the benefit of the community, it should be made clear that this is the case. And it should be easy to guide users to changing the default curve, which is why a link is provided to jump to the appropriate section.