Exit sender when room matches but code doesn't
Previously if the receiver introduced a code that would hash to an open
room, but the code would be different from the code used to generate the
room, the receiver would fail. However, the sender would keep running,
in spite the fact that subsequent attempts to connect to the room would
fail even with the correct password. This change ensures the receiver
also stops.

- Some errors were also refactored to make the error checks easier.