Abort sending when message authentication fails (#980)
This either means that the TCP connection did not use the proper key to
encrypt (which should never happen when running properly) or the receiver
wrote a key which starts with the same 4 characters, but does not match the
sender's code. If the latter, regardless if the user attempts to read the
message by correcting the key, it will always receive an error after the first
failure. However, the sender will not be closed, making it difficult to detect
what happened and why the transfer does not succeed. This change also forces the
sender to close when the receiver uses a key that starts with the same characters,
but is a different one